requirements.txt
fastapi
SQLAlchemy
asyncpg
pydantic
uvicorn
anyio
starlette
pydantic_core
typing_extensions
click
colorama
greenlet
h11
idna
sniffio
annotated-types
alembic
python-multipart
httpx
pytest
pytest-asyncio
PyJWT